chore: 🐋 SelectedList 관련 types 파일 분리
